+++ integration/presets/elements.ts
@@ -2,9 +2,6 @@ import { constants } from '../constants';
 import { applicationConfig } from '../models/applicationConfig.js';
 import { templates } from '../templates/index.js';
 
-const clerkNextjsLocal = `file:${process.cwd()}/packages/nextjs`;
-const clerkElementsLocal = `file:${process.cwd()}/packages/elements`;
-
 const nextAppRouter = applicationConfig()
   .setName('elements-next')
   .useTemplate(templates['elements-next'])
@@ -16,8 +13,8 @@ const nextAppRouter = applicationConfig()
   .addDependency('next', constants.E2E_NEXTJS_VERSION)
   .addDependency('react', constants.E2E_REACT_VERSION)
   .addDependency('react-dom', constants.E2E_REACT_DOM_VERSION)
-  .addDependency('@clerk/nextjs', constants.E2E_CLERK_VERSION || clerkNextjsLocal)
-  .addDependency('@clerk/elements', constants.E2E_CLERK_VERSION || clerkElementsLocal);
+  .addDependency('@clerk/nextjs', constants.E2E_CLERK_VERSION || '*')
+  .addDependency('@clerk/elements', constants.E2E_CLERK_VERSION || '*');
 
 export const elements = {
   nextAppRouter,