+++ integration/presets/expo.ts
@@ -1,8 +1,6 @@
 import { applicationConfig } from '../models/applicationConfig';
 import { templates } from '../templates';
 
-const clerkExpoLocal = `file:${process.cwd()}/packages/expo`;
-
 const expoWeb = applicationConfig()
   .setName('expo-web')
   .useTemplate(templates['expo-web'])
@@ -11,7 +9,7 @@ const expoWeb = applicationConfig()
   .addScript('dev', 'pnpm dev')
   .addScript('build', 'pnpm build')
   .addScript('serve', 'pnpm start')
-  .addDependency('@clerk/clerk-expo', clerkExpoLocal);
+  .addDependency('@clerk/clerk-expo', '*');
 
 export const expo = {
   expoWeb,