+++ integration/presets/next.ts
@@ -2,9 +2,6 @@ import { constants } from '../constants';
 import { applicationConfig } from '../models/applicationConfig.js';
 import { templates } from '../templates/index.js';
 
-const clerkNextjsLocal = `file:${process.cwd()}/packages/nextjs`;
-const clerkSharedLocal = `file:${process.cwd()}/packages/shared`;
-const clerkTypesLocal = `file:${process.cwd()}/packages/types`;
 const appRouter = applicationConfig()
   .setName('next-app-router')
   .useTemplate(templates['next-app-router'])
@@ -16,9 +13,9 @@ const appRouter = applicationConfig()
   .addDependency('next', constants.E2E_NEXTJS_VERSION)
   .addDependency('react', constants.E2E_REACT_VERSION)
   .addDependency('react-dom', constants.E2E_REACT_DOM_VERSION)
-  .addDependency('@clerk/nextjs', constants.E2E_CLERK_VERSION || clerkNextjsLocal)
-  .addDependency('@clerk/shared', clerkSharedLocal)
-  .addDependency('@clerk/types', clerkTypesLocal);
+  .addDependency('@clerk/nextjs', constants.E2E_CLERK_VERSION || '*')
+  .addDependency('@clerk/shared', '*')
+  .addDependency('@clerk/types', '*');
 
 const appRouterTurbo = appRouter.clone().setName('next-app-router-turbopack').addScript('dev', 'pnpm dev');